Add custom class components for javascript to use
The nav class captures the whole navigation group that is set up for
mobile navigation. A hamburger is also created using before and after
elements to create the lines.
Both nav and hamburger have classes that will toggle when the hamburger
is clicked on. That logic will be written in a seprate js file.