Finish mobile navigation
All logic and classes for mobile navigation are fully implemented by
adding the index.js file to the script tag in the head. The tailwindcss
compenents are added to the correct elements and the hamburger div is
added.
Any bugs or styling changes for other screen sizes were also made.